generalized TimePeriodRelation entity
